+++ backend/tsconfig.json
@@ -21,5 +21,5 @@
     "outDir": "./dist",
     "sourceRoot": "/src"
   },
-  "exclude": ["node_modules", "dist", "scripts", "__tests__"]
+  "exclude": ["node_modules", "dist", "scripts", "src/__tests__"]
 }
